CCD-3086 Add JDK 17 to BEFTA Framework - Gradle upgraded from 4.10.3 to 7.4.2

+++ src/test/java/uk/gov/hmcts/befta/TestAutomationConfigTest.java
@@ -218,7 +218,7 @@ class TestAutomationConfigTest {
             value = BEFTA_USER_TOKEN_CACHE_TTL_VALUE)
     void testGetUserTokenCacheTtlInSeconds() {
         assertEquals(Long.parseLong(BEFTA_USER_TOKEN_CACHE_TTL_VALUE),
-                TestAutomationConfig.INSTANCE.getUserTokenCacheTtlInSeconds());
+                TestAutomationConfig.INSTANCE.getUserTokenCacheTtlInSeconds().longValue());
     }
 
     /**
@@ -228,7 +228,7 @@ class TestAutomationConfigTest {
     @Test
     void testGetUserTokenCacheTtlInSecondsNoValueSet() {
         assertEquals(0L,
-                TestAutomationConfig.INSTANCE.getUserTokenCacheTtlInSeconds());
+                TestAutomationConfig.INSTANCE.getUserTokenCacheTtlInSeconds().longValue());
     }
 
     /**
@@ -240,7 +240,7 @@ class TestAutomationConfigTest {
             value = BEFTA_S2S_TOKEN_CACHE_TTL_VALUE)
     void testGetS2STokenCacheTtlInSeconds() {
         assertEquals(Long.parseLong(BEFTA_S2S_TOKEN_CACHE_TTL_VALUE),
-                TestAutomationConfig.INSTANCE.getS2STokenCacheTtlInSeconds());
+                TestAutomationConfig.INSTANCE.getS2STokenCacheTtlInSeconds().longValue());
     }
 
     /**
@@ -250,7 +250,7 @@ class TestAutomationConfigTest {
     @Test
     void testGetS2STokenCacheTtlInSecondsNoValuesSet() {
         assertEquals(0L,
-                TestAutomationConfig.INSTANCE.getS2STokenCacheTtlInSeconds());
+                TestAutomationConfig.INSTANCE.getS2STokenCacheTtlInSeconds().longValue());
     }
 
     /**
